moved commands into their own files
