+++ src/index.js
@@ -1,3 +1,5 @@
+// OOP design highly inspired by https://github.com/MenuDocs/discord.js-template
+
 import 'dotenv/config.js';
 import MDClient from './Structures/MDClient.js';
